Move ModelProcessor and TableProcessor out of Moco.
