+++ Bindings/OpenSimHeaders_moco.h
@@ -4,7 +4,6 @@
  */
 
 #include <OpenSim/Moco/About.h>
-#include <OpenSim/Moco/Common/TableProcessor.h>
 #include <OpenSim/Moco/Components/DeGrooteFregly2016Muscle.h>
 #include <OpenSim/Moco/Components/MultivariatePolynomialFunction.h>
 #include <OpenSim/Moco/Components/PositionMotion.h>
@@ -12,6 +11,7 @@
 #include <OpenSim/Moco/MocoCasADiSolver/MocoCasADiSolver.h>
 #include <OpenSim/Moco/MocoControlBoundConstraint.h>
 #include <OpenSim/Moco/MocoFrameDistanceConstraint.h>
+#include <OpenSim/Moco/MocoGoal/MocoAccelerationTrackingGoal.h>
 #include <OpenSim/Moco/MocoGoal/MocoContactTrackingGoal.h>
 #include <OpenSim/Moco/MocoGoal/MocoControlGoal.h>
 #include <OpenSim/Moco/MocoGoal/MocoControlTrackingGoal.h>
@@ -23,7 +23,6 @@
 #include <OpenSim/Moco/MocoGoal/MocoMarkerFinalGoal.h>
 #include <OpenSim/Moco/MocoGoal/MocoMarkerTrackingGoal.h>
 #include <OpenSim/Moco/MocoGoal/MocoOrientationTrackingGoal.h>
-#include <OpenSim/Moco/MocoGoal/MocoAccelerationTrackingGoal.h>
 #include <OpenSim/Moco/MocoGoal/MocoOutputGoal.h>
 #include <OpenSim/Moco/MocoGoal/MocoPeriodicityGoal.h>
 #include <OpenSim/Moco/MocoGoal/MocoStateTrackingGoal.h>
@@ -39,7 +38,7 @@
 #include <OpenSim/Moco/MocoTropterSolver.h>
 #include <OpenSim/Moco/MocoUtilities.h>
 #include <OpenSim/Moco/MocoWeightSet.h>
-#include <OpenSim/Moco/ModelOperators.h>
+#include <OpenSim/Moco/ModelOperatorsDGF.h>
 #include <OpenSim/Moco/osimMocoDLL.h>
 
 #endif // OPENSIM_OPENSIM_HEADERS_MOCO_H_
