+++ OpenSim/Actuators/RegisterTypes_osimActuators.cpp
@@ -58,6 +58,8 @@
 #include "Millard2012EquilibriumMuscle.h"
 #include "Millard2012AccelerationMuscle.h"
 
+#include "ModelOperators.h"
+
 // Awaiting new component architecture that supports subcomponents with states.
 //#include "ConstantMuscleActivation.h"
 //#include "ZerothOrderMuscleActivationDynamics.h"
@@ -106,6 +108,15 @@ OSIMACTUATORS_API void RegisterTypes_osimActuators()
 
     Object::RegisterType(Millard2012EquilibriumMuscle());
     Object::RegisterType(Millard2012AccelerationMuscle());
+
+    Object::registerType(ModelProcessor());
+    Object::registerType(ModOpIgnoreActivationDynamics());
+    Object::registerType(ModOpIgnoreTendonCompliance());
+    Object::registerType(ModOpScaleMaxIsometricForce());
+    Object::registerType(ModOpRemoveMuscles());
+    Object::registerType(ModOpAddReserves());
+    Object::registerType(ModOpAddExternalLoads());
+    Object::registerType(ModOpReplaceJointsWithWelds());
 
     //Object::RegisterType( ConstantMuscleActivation() );
     //Object::RegisterType( ZerothOrderMuscleActivationDynamics() );
