+++ OpenSim/Actuators/Test/testModelProcessor.cpp
@@ -17,9 +17,10 @@
  * -------------------------------------------------------------------------- */
 
 #define CATCH_CONFIG_MAIN
-#include "Testing.h"
+#include "OpenSim/Moco/Test/Testing.h"
 
 #include <OpenSim/Actuators/Millard2012EquilibriumMuscle.h>
+#include <OpenSim/Actuators/ModelOperators.h>
 #include <OpenSim/Analyses/MuscleAnalysis.h>
 #include <OpenSim/Moco/osimMoco.h>
 #include <OpenSim/Simulation/SimbodyEngine/PinJoint.h>
