+++ OpenSim/Examples/Moco/example3DWalking/exampleMocoTrack.cpp
@@ -25,8 +25,9 @@
 ///
 /// See the README.txt next to this file for more information.
 
-#include <OpenSim/Moco/osimMoco.h>
 #include <OpenSim/Actuators/CoordinateActuator.h>
+#include <OpenSim/Actuators/ModelOperators.h>
+#include <OpenSim/Moco/osimMoco.h>
 
 using namespace OpenSim;
