+++ OpenSim/Moco/Test/testMocoTrack.cpp
@@ -16,6 +16,7 @@
  * limitations under the License.                                             *
  * -------------------------------------------------------------------------- */
 
+#include <OpenSim/Actuators/ModelOperators.h>
 #include <OpenSim/Moco/osimMoco.h>
 
 #define CATCH_CONFIG_MAIN
