+++ OpenSim/Simulation/RegisterTypes_osimSimulation.cpp
@@ -125,6 +125,7 @@
 #include "OpenSense/IMUPlacer.h"
 #include "OrientationsReference.h"
 #include "StatesTrajectoryReporter.h"
+#include "TableProcessor.h"
 
 #include <string>
 #include <iostream>
@@ -267,6 +268,11 @@ OSIMSIMULATION_API void RegisterTypes_osimSimulation()
     Object::registerType( IMUPlacer() );
     
     Object::registerType( StatesTrajectoryReporter() );
+
+    Object::registerType( TableProcessor() );
+
+    Object::registerType( TabOpLowPassFilter() );
+    Object::registerType( TabOpUseAbsoluteStateNames() );
 
     // OLD Versions
     // Associate an instance with old name to help deserialization.
